fix: rendering of 0xAD (soft hyphen) with lualatex is broken
the font contains 0xAD, which represents the soft hyphen.  pillow
can use this codepoint without problems, but lualatex produces
incorrect output when it renders 0xAD.  there seem to be many
pitfalls when one wants to render a soft hyphen correctly, so the
font designers decided to offer the same image under another code
point, 0xFE, and we just use that.